system1/requirements.txt
python-dotenv
setuptools
pydantic
numpy
uvicorn
mcp
fastapi
fastapi-mcp
langchain
langchain-text-splitters
langchain-community
langchain-ollama
langchain-milvus
langchain-mongodb
langchain-neo4j
llama-index
llama-index-core
llama-index-experimental
llama-index-llms-ollama
llama-index-llms-huggingface
llama-index-embeddings-ollama
llama-index-embeddings-langchain
llama-index-readers-database
llama-index-readers-file
llama-index-readers-json
llama-index-readers-milvus
llama-index-readers-mongodb
llama-index-readers-obsidian
llama-index-readers-whisper
llama-index-readers-graphdb-cypher
llama-index-vector-stores-milvus
llama-index-extractors-entity
llama-index-storage-index-store-elasticsearch
llama-index-storage-index-store-mongodb
llama-index-storage-index-store-redis
llama-index-storage-docstore-elasticsearch
llama-index-storage-docstore-mongodb
llama-index-storage-docstore-redis
llama-index-storage-kvstore-elasticsearch
llama-index-storage-kvstore-mongodb
llama-index-storage-kvstore-redis
llama-index-tools-requests
llama-index-tools-mcp
llama-index-tools-neo4j
llama-index-tools-database
llama-index-graph-stores-neo4j
llama-index-packs-neo4j-query-engine
llama-index-multi-modal-llms-ollama
transformers
sentence-transformers
ollama
pymilvus
pymongo
redis
httpx
aiohttp
tiktoken
torch
sqlalchemy
pymssql
jinja2
pika
neo4j
numpy
pandas
yomitoku
Pillow
PyMuPDF
opencv-python
pytesseract
ebooklib
yomitoku
markdown
chardet